requirements.txt
fastapi==0.115.8
uvicorn==0.34.0

pydantic[email]
sqlalchemy==2.0.38
sqlmodel==0.0.22
alembic==1.14.1

PyJWT==2.10.1
passlib[bcrypt]

dotenv==0.9.9
